norminetted everything
