+++ sort_cmd_utils.c
@@ -3,18 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   sort_cmd_utils.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: anasr <anasr@student.42.fr>                +#+  +:+       +#+        */
+/*   By: aalsuwai <aalsuwai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/04 16:16:47 by anasr             #+#    #+#             */
-/*   Updated: 2022/04/04 18:05:00 by anasr            ###   ########.fr       */
+/*   Updated: 2022/04/05 13:35:32 by aalsuwai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-void	sort_input(int array_index, char **specific_cmd, t_sort_indices *index, t_parser_info *p)
+void	sort_input(int array_index, char **specific_cmd, \
+t_sort_indices *index, t_parser_info *p)
 {
-	char *temp;
+	char	*temp;
 
 	if (!ft_strncmp(specific_cmd[index->i], "<<", 2))
 		p->in_arrow_flag[array_index][index->in_index] = DOUBLE_ARROW;
@@ -26,18 +27,21 @@ void	sort_input(int array_index, char **specific_cmd, t_sort_indices *index, t_p
 		temp = expand_dollars_in_str(specific_cmd[index->i + 1], p, false);
 		if (!temp[0])
 		{
-			printf("babyshell: %s: ambiguous redirect\n", specific_cmd[index->i + 1]);
+			printf("babyshell: %s: ambiguous redirect\n", \
+			specific_cmd[index->i + 1]);
 			p->in_arrow_flag[array_index][index->in_index] = -1;
 		}
 		free(specific_cmd[index->i + 1]);
 		specific_cmd[index->i + 1] = temp;
 	}
-	p->input_files_delimiters[array_index][index->in_index++] = specific_cmd[++index->i];
+	p->input_files_delimiters[array_index][index->in_index++] = \
+	specific_cmd[++index->i];
 }
 
-void	sort_output(int array_index, char **specific_cmd, t_sort_indices *index, t_parser_info *p)
+void	sort_output(int array_index, char **specific_cmd, \
+t_sort_indices *index, t_parser_info *p)
 {
-	char *temp;
+	char	*temp;
 
 	if (!ft_strncmp(specific_cmd[index->i], ">>", 2))
 		p->out_arrow_flag[array_index][index->out_index] = DOUBLE_ARROW;
@@ -48,7 +52,8 @@ void	sort_output(int array_index, char **specific_cmd, t_sort_indices *index, t_
 		temp = expand_dollars_in_str(specific_cmd[index->i + 1], p, false);
 		if (!temp[0])
 		{
-			printf("babyshell: %s: ambiguous redirect\n", specific_cmd[index->i + 1]);
+			printf("babyshell: %s: ambiguous redirect\n", \
+			specific_cmd[index->i + 1]);
 			p->out_arrow_flag[array_index][index->out_index] = -1;
 		}
 		free(specific_cmd[index->i + 1]);
@@ -57,8 +62,8 @@ void	sort_output(int array_index, char **specific_cmd, t_sort_indices *index, t_
 	p->output_files[array_index][index->out_index++] = specific_cmd[++index->i];
 }
 
-
-void	sort_cmds(int array_index, char **specific_cmd, t_sort_indices *index, t_parser_info *p)
+void	sort_cmds(int array_index, char **specific_cmd, \
+t_sort_indices *index, t_parser_info *p)
 {
 	char	*temp;
 
@@ -70,12 +75,14 @@ void	sort_cmds(int array_index, char **specific_cmd, t_sort_indices *index, t_pa
 		free(specific_cmd[index->i]);
 		specific_cmd[index->i] = temp;
 	}
-	if (ft_strchr(specific_cmd[index->i], '\'') || ft_strchr(specific_cmd[index->i], '\"'))
+	if (ft_strchr(specific_cmd[index->i], '\'') \
+	|| ft_strchr(specific_cmd[index->i], '\"'))
 		specific_cmd[index->i] = strcpy_wout_quotes(specific_cmd[index->i]);
 	p->cmd[array_index][index->cmd_index++] = specific_cmd[index->i];
 }
 
-void	save_cmds_helper(int array_index, char *input, char **meta, t_parser_info *p)
+void	save_cmds_helper(int array_index, char *input, char **meta, \
+t_parser_info *p)
 {
 	allocate_meme_specific(input, array_index, p);
 	p->cmd_part[array_index] = ft_split_custom(input, meta, p);
